Add full consumer/producer props to binder/binding props
* Re-use the consumer and producer props from PulsarProperties
* Add unit and integration test for newly added props
