+++ spring-pulsar-spring-cloud-stream-binder/src/test/java/org/springframework/pulsar/spring/cloud/stream/binder/PulsarMessageChannelBinderResolveSchemaTests.java
@@ -37,6 +37,7 @@ import org.springframework.pulsar.core.PulsarConsumerFactory;
 import org.springframework.pulsar.core.PulsarTemplate;
 import org.springframework.pulsar.core.Resolved;
 import org.springframework.pulsar.core.SchemaResolver;
+import org.springframework.pulsar.spring.cloud.stream.binder.properties.PulsarBinderConfigurationProperties;
 import org.springframework.pulsar.spring.cloud.stream.binder.provisioning.PulsarTopicProvisioner;
 
 /**
@@ -50,7 +51,8 @@ public class PulsarMessageChannelBinderResolveSchemaTests {
 
 	@SuppressWarnings("unchecked")
 	private PulsarMessageChannelBinder binder = new PulsarMessageChannelBinder(mock(PulsarTopicProvisioner.class),
-			mock(PulsarTemplate.class), mock(PulsarConsumerFactory.class), resolver);
+			mock(PulsarTemplate.class), mock(PulsarConsumerFactory.class),
+			mock(PulsarBinderConfigurationProperties.class), resolver);
 
 	@ParameterizedTest
 	@EnumSource(mode = Mode.MATCH_NONE, names = "^(AUTO.*|AVRO|JSON|KEY_VALUE|NONE|PROTOBUF.*)$")
